authentication

+++ server/src/routes/memories.ts
@@ -3,10 +3,17 @@ import { FastifyInstance } from "fastify";
 import { z } from "zod";
 
 export async function memoriesRoute(app: FastifyInstance) {
+  app.addHook("preHandler", async (request) => {
+    await request.jwtVerify();
+  });
+
   const prisma = new PrismaClient();
 
-  app.get("/memories", async () => {
+  app.get("/memories", async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createDate: "asc",
       },
@@ -20,7 +27,7 @@ export async function memoriesRoute(app: FastifyInstance) {
     });
   });
 
-  app.get("/memories/:id", async (request) => {
+  app.get("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -32,6 +39,10 @@ export async function memoriesRoute(app: FastifyInstance) {
         id,
       },
     });
+
+    if (!memory.isPublic && request.user.sub !== memory.userId) {
+      return reply.status(401).send();
+    }
     return memory;
   });
 
@@ -49,12 +60,12 @@ export async function memoriesRoute(app: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: "asdk",
+        userId: request.user.sub,
       },
     });
   });
 
-  app.put("/memories/:id", async (request) => {
+  app.put("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
@@ -68,6 +79,16 @@ export async function memoriesRoute(app: FastifyInstance) {
 
     const { content, isPublic, coverUrl } = bodySchema.parse(request.body);
 
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
+
     await prisma.memory.update({
       where: {
         id,
@@ -80,12 +101,22 @@ export async function memoriesRoute(app: FastifyInstance) {
     });
   });
 
-  app.delete("/memories/:id", async (request) => {
+  app.delete("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     });
 
     const { id } = paramsSchema.parse(request.params);
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send();
+    }
 
     await prisma.memory.delete({
       where: {
